==> 10_cutoff_survival_genes.ipynb <==
import pandas as pd

mdel_pgain = pd.read_csv('survival_results/survival_mdel_pgain_tcga_glioma.csv', index_col = 0)
mdel_pgain

# keep rows with p-value < 0.01
mdel_pgain = mdel_pgain[mdel_pgain['p_value'] < 0.01]

# keep rows with long_survival == low
mdel_pgain = mdel_pgain[mdel_pgain['long_survival'] == 'low']

mdel_pgain_list = mdel_pgain.index.tolist()

# save the list to a file
with open('survival_mdel_pgain_tcga_glioma_list.txt', 'w') as f:
    for item in mdel_pgain_list:
        f.write("%s\n" % item)

---

genes = pd.read_csv('survival_mdel_pgain_tcga_glioma_list.txt', sep = ": ", index_col = 0, header = None)

genes

genes_secretory = genes[genes[1] == 'Secretory']
genes_surface = genes[genes[1] == 'Surface']
genes_cytosolic = genes[genes[1] == 'Cytosolic']
genes_unclear = genes[genes[1] == '모름']

genes_surface
